Refactoring: Move service generation into separate file.
